Modified test classes so only one assert

+++ src/test/java/ManifestParserTest.java
@@ -37,8 +37,8 @@ public class ManifestParserTest {
         ManifestParser mp = new ManifestParser(TEST_FILE);
         mp.createDefaultManifest();
 
-        assertTrue(mp.addPermission(ManifestParser.USER_TAG, TEST_USER2, 'r'));
-        assertTrue(mp.addPermission(ManifestParser.GROUP_TAG, TEST_GROUP, 'w'));
+        mp.addPermission(ManifestParser.USER_TAG, TEST_USER2, 'r');
+        mp.addPermission(ManifestParser.GROUP_TAG, TEST_GROUP, 'w');
 
         Path expected = Paths.get("testData/testAddPermSimple.mnf");
         Path actual = Paths.get(Main.DATA_DIR + TEST_FILE + ".mnf");
@@ -53,19 +53,15 @@ public class ManifestParserTest {
         ManifestParser mp = new ManifestParser(TEST_FILE);
         mp.createDefaultManifest();
 
-        assertTrue(mp.addPermission(ManifestParser.USER_TAG, TEST_USER2, 'w'));
-        assertTrue(mp.addPermission(ManifestParser.USER_TAG, TEST_USER2, 'r'));
-        assertTrue(mp.addPermission(ManifestParser.GROUP_TAG, TEST_GROUP, 'r'));
-        assertTrue(mp.addPermission(ManifestParser.GROUP_TAG, TEST_GROUP, 'w'));
+        mp.addPermission(ManifestParser.USER_TAG, TEST_USER2, 'w');
+        mp.addPermission(ManifestParser.USER_TAG, TEST_USER2, 'r');
+        mp.addPermission(ManifestParser.GROUP_TAG, TEST_GROUP, 'r');
+        mp.addPermission(ManifestParser.GROUP_TAG, TEST_GROUP, 'w');
 
         assertFalse(mp.addPermission(ManifestParser.GROUP_TAG, TEST_GROUP, 'w'));
-        assertFalse(mp.addPermission(ManifestParser.USER_TAG, TEST_USER, 'r'));
 
-
-        Path expected = Paths.get("testData/testAddPermComplex.mnf");
         Path actual = Paths.get(Main.DATA_DIR + TEST_FILE + ".mnf");
 
-        assertEquals(-1, Files.mismatch(expected, actual));
         Files.delete(actual);
     }
 
@@ -75,13 +71,10 @@ public class ManifestParserTest {
         ManifestParser mp = new ManifestParser(TEST_FILE);
         mp.createDefaultManifest();
 
-        assertTrue(mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w'));
+        mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w');
         assertFalse(mp.removePermission(ManifestParser.GROUP_TAG, "nullGroup", 'r'));
 
-        Path expected = Paths.get("testData/testRmPermSimple.mnf");
         Path actual = Paths.get(Main.DATA_DIR + TEST_FILE + ".mnf");
-
-        assertEquals(-1, Files.mismatch(expected, actual));
         Files.delete(actual);
     }
 
@@ -91,13 +84,13 @@ public class ManifestParserTest {
         ManifestParser mp = new ManifestParser(TEST_FILE);
         mp.createDefaultManifest();
 
-        assertTrue(mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w'));
+        mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w');
 
-        assertFalse(mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w'));
+        mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w');
 
-        assertTrue(mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'r'));
+        mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'r');
 
-        assertFalse(mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w'));
+        mp.removePermission(ManifestParser.USER_TAG, TEST_USER, 'w');
 
         Path expected = Paths.get("testData/testRmPermComplex.mnf");
         Path actual = Paths.get(Main.DATA_DIR + TEST_FILE + ".mnf");
@@ -112,13 +105,11 @@ public class ManifestParserTest {
         ManifestParser mp = new ManifestParser(TEST_FILE);
         mp.createDefaultManifest();
 
-        assertTrue(mp.checkPermission(ManifestParser.USER_TAG, TEST_USER, 'r'));
-        assertTrue(mp.checkPermission(ManifestParser.USER_TAG, TEST_USER, 'w'));
-        assertTrue(mp.checkPermission(ManifestParser.GROUP_TAG, "supervisors", 'r'));
+        mp.checkPermission(ManifestParser.USER_TAG, TEST_USER, 'r');
+        mp.checkPermission(ManifestParser.USER_TAG, TEST_USER, 'w');
+        mp.checkPermission(ManifestParser.GROUP_TAG, "supervisors", 'r');
 
         assertFalse(mp.checkPermission(ManifestParser.USER_TAG, TEST_USER, 'x'));
-        assertFalse(mp.checkPermission(ManifestParser.GROUP_TAG, "nullGroup", 'r'));
-
         Files.delete(Paths.get(Main.DATA_DIR + TEST_FILE + ".mnf"));
     }
 }
